+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Undeclared variable diagnostics: allow @Set <VarName> outside of function to not display error if it is declared
 
 ## [0.0.3] - 2026-01-12
+### Added
+- Added documentation for @Get, @Set
 ### Fixed
 - Variables used as function parameters to not show error
 - fixed '' in If